更新

+++ src/app/library/worker/worker.service.ts
@@ -71,44 +71,57 @@ export class WorkerService {
     const res = await this.UrlToBolbUrl('http://localhost:7700/bilibili/comics/26574');
   }
 
-  workerImageCompression(urls, width, quality) {
-    return new Promise(async (resolve, reject) => {
-      const res = await Promise.all(urls.map(x => `${x}?width=${width}&quality=${quality}`).map(x => this.webCh.match('image', x)))
+  async workerImageCompression(urls, width, quality) {
+    if(!urls) return []
+    if(urls&&urls.length==0) return []
+    else if(urls[0].substring(7, 21) == "localhost:7700"){
+      return new Promise(async (resolve, reject) => {
+        const res = await Promise.all(urls.map(x => `${x}?width=${width}&quality=${quality}`).map(x => this.webCh.match('image', x)))
 
-      if (res) {
-        let num = 0;
-        for (let index = 0; index < res.length; index++) {
-          const element = res[index];
-          if (element) num++;
-        }
-        if (num == urls.length) resolve(urls.map(x => `${x}?width=${width}&quality=${quality}`))
-      }
-
-      function chunkArray<T>(array: T[], chunkSize: number): T[][] {
-        const chunks: T[][] = [];
-        for (let i = 0; i < array.length; i += chunkSize) {
-          chunks.push(array.slice(i, i + chunkSize));
-        }
-        return chunks;
-      }
-
-
-      let max = navigator.hardwareConcurrency * 2;
-      let num = 0;
-      for (let index = 0; index < max; index++) {
-        const worker = new Worker(new URL('./app.worker', import.meta.url));
-        worker.postMessage({ type: "compress_multiple", data: chunkArray(urls, Math.ceil(urls.length / max))[index], width, quality })
-        worker.onmessage = async ({ data }) => {
-          if (data.type == "destroy") {
-            num++
-            worker.terminate();
-            if (num == max) {
-              resolve(urls.map(x => `${x}?width=${width}&quality=${quality}`))
-            }
+        if (res) {
+          let num = 0;
+          for (let index = 0; index < res.length; index++) {
+            const element = res[index];
+            if (element) num++;
           }
-        };
+          if (num == urls.length) resolve(urls.map(x => `${x}?width=${width}&quality=${quality}`))
+        }
+
+        function chunkArray<T>(array: T[], chunkSize: number): T[][] {
+          const chunks: T[][] = [];
+          for (let i = 0; i < array.length; i += chunkSize) {
+            chunks.push(array.slice(i, i + chunkSize));
+          }
+          return chunks;
+        }
+
+
+        let max = navigator.hardwareConcurrency * 2;
+        let num = 0;
+        for (let index = 0; index < max; index++) {
+          const worker = new Worker(new URL('./app.worker', import.meta.url));
+          worker.postMessage({ type: "compress_multiple", data: chunkArray(urls, Math.ceil(urls.length / max))[index], width, quality })
+          worker.onmessage = async ({ data }) => {
+            if (data.type == "destroy") {
+              num++
+              worker.terminate();
+              if (num == max) {
+                resolve(urls.map(x => `${x}?width=${width}&quality=${quality}`))
+              }
+            }
+          };
+        }
+      })
+    }else{
+      let arr=[];
+      for (let index = 0; index < urls.length; index++) {
+        const blob = await this.Image.getImageBlob(urls[index]);
+        const url=URL.createObjectURL(blob)
+        arr.push(url)
       }
-    })
+      return arr
+    }
+
 
   }
